fix(atc): use Convert.ToInt64 in EnumHelper flag-skip logic; add long.IsBinarySequence
ShouldEnumValueBeSkipped used Convert.ToInt32 to read boxed enum values,
throwing OverflowException for [Flags] enums whose backing type is long
and whose values exceed int.MaxValue.

Changed all three ToInt32 calls in that private helper to ToInt64 and
changed the local variable type from int to long. Added IsBinarySequence
as a long extension to support the flag-bit check on the wider type.

+++ src/Atc/Helpers/Enums/EnumHelper.cs
@@ -636,7 +636,7 @@ public static class EnumHelper
         bool byFlagIncludeBase,
         bool byFlagIncludeCombined)
     {
-        if (!includeDefault && Convert.ToInt32(objEnumValue, CultureInfo.InvariantCulture) == 0)
+        if (!includeDefault && Convert.ToInt64(objEnumValue, CultureInfo.InvariantCulture) == 0L)
         {
             return true;
         }
@@ -646,7 +646,7 @@ public static class EnumHelper
             return false;
         }
 
-        var n = Convert.ToInt32(objEnumValue, CultureInfo.InvariantCulture);
+        var n = Convert.ToInt64(objEnumValue, CultureInfo.InvariantCulture);
         if (!byFlagIncludeBase && n.IsBinarySequence())
         {
             return true;
@@ -662,6 +662,6 @@ public static class EnumHelper
             return false;
         }
 
-        return !includeDefault || Convert.ToInt32(objEnumValue, CultureInfo.InvariantCulture) != 0;
+        return !includeDefault || Convert.ToInt64(objEnumValue, CultureInfo.InvariantCulture) != 0L;
     }
 }
